refactor(maps): split LocationResolver tests into focused files
Split 598-line test file into 3 self-contained files:
- LocationResolver.unit.spec.ts (247 lines) - Basic functionality, empty inputs, API keys
- LocationResolver.integration.spec.ts (321 lines) - Mixed markers, caching, errors, integration
- LocationResolver.reactivity.spec.ts (226 lines) - MobX reactions and observable behavior

Benefits:
- Each file self-contained with inline setup (no helpers folder)
- Follows Gallery/Datagrid patterns in the repo
- Easy to locate specific test types
- Can run test files independently
- No abstraction layers to learn
- Clear test intent without jumping to other files

All 45 tests passing with 100% model layer coverage maintained
